started working on step5 (matching)

+++ step4/Learner.py
@@ -11,7 +11,7 @@ class Learner:
         self.collected_rewards = []
         self.m = np.array([config.TOT_CUSTOMERS // 4 for _ in range(4)])  # Non-informative prior
         self.s_2 = np.array([1.0, 1.0, 1.0, 1.0])
-        self.expected_customers = np.random.normal(self.m, np.sqrt(self.s_2))  # initial number of expected customers per class, according to our prior
+        self.expected_customers = np.random.normal(self.m, np.sqrt(self.s_2)).astype(int)  # initial number of expected customers per class, according to our prior
 
     def update_observations(self, pulled_arm, reward):
         self.rewards_per_arm[pulled_arm].append(reward)  # ... (for UCB ???)
@@ -24,5 +24,5 @@ class Learner:
 
         self.m = (s_2 * x_bar + m * sigma_2) / (s_2 + sigma_2)
         self.s_2 = (s_2 * sigma_2) / (s_2 + sigma_2)
-        self.expected_customers = np.random.normal(self.m, np.sqrt(self.s_2))
+        self.expected_customers = np.random.normal(self.m, np.sqrt(self.s_2)).astype(int)
